Exposed logitic regression functionality
Created two ways of doing logistic regression. Either considering each line in the data seperately or first preprocessing the data to count the number of pools of each size and each result for each group. The former is simpler and more robust, but slower. Future development will focus on this.