Change default prior for group effect sd in HierPoolPrev
Allow users to set (hyper) prior on random/group effects in for use in the HeirPoolPrev function. The prior is half Cauchy(0, S) where S is set to 2 by default. Previously the default was 25, which was too diffuse in typical setups, and users were unable to change the prior.

Also switches from using alpha level for confidence to confidence level, for functions PoolPrev and HierPoolPrev. This is now consistent with getPrevalence