HierPoolPrev population prevalence correction
Previously we were estimating population prevalence as the logit(mu) where mu was the intercept on the logit scale. However this isn't correct. Prevalence is actually the mean of a logitnormal distribution with parameters mu and sigma^2 (where sigma^2 is the sum of the random-effect variances). I now do this properly. However to keep defining the prior for prevalence stan needs to calculate two integrals at every step: one to calculate the mean of the logitnormal distribution and a second to correct for the fact the the prior is on the prevalence while the sampling is happening on mu. This solution works and maintains the same interface, but is extremely slow.