Remove bias from HierPoolPrev
This has also required changing how priors are set. Priors are now analogous to PoolRegBayes, with priors on intercept and group-effect standard deviations. Since what we are actually interested in is the total variation across all grouping variables, and to avoid the prior for variation increasing with group size, the default model sets the prior on the root-sum-square of group-effect sd, however there is an option for setting priors on each sd. This different parameters are being sampled in each setup this requires two separate .stan files.

Therefore I have renamed all (now 3) .stan files to more informative