Updating getPrevalence generic for "glmerMod"
Lots of commenting to explain my way through the process.

To fix the R CMD CHECK issue, the key was to add the PredDataSub$..allnegative column to the Prev tibble so that the ..zeroest column could be calculated. I then remove the PredDataSub$..allnegative column to prevent duplicates when colbinding PredDataSub with Prev.