Add utility function to return incorrectly nested variables
For a given inner and outer cluster name (where the outer cluster is larger in the hierarchy), the function `check_nesting_levels()` looks at each variable in the inner cluster column and checks how many unique outer clusters contain that value. If the value is included in multiple output clusters, the different outer cluster values for the given inner cluster value are recorded and output.