CheckInputData() only adds column for bad hierarchy
CheckInputData() now has a tryCatch statement where it attempts to call CheckClusterVars(). If the nesting warning is raised, then a new column PoolTestR ID as added to the dataframe which has a unique ID for each location. If any other errors or warnings are raised then CheckInputData() prints the error/message to the user and returns NA. If data with good nesting/hierarchy is input, the dataframe is returned unchanged.